fix: rename test_settings to test_core and fix stale settings imports
- Rename tests/events/test_settings.py → test_core.py (docstring + import)
- Rename tests/mappers/test_settings.py → test_core.py (docstring + typing import + import)
- Rename tests/repos/test_settings.py → test_core.py (docstring + import)
- Fix stale 'from tiferet.events.settings' imports in 13 other test files
- Fix stale 'from tiferet.interfaces.settings' import in test_middleware.py
- Fix stale 'from ..events.settings' imports in tiferet/utils/{yaml,json,toml}.py

All 756 tests pass.

Addresses review findings 1a, 1b, 1c on PR #919.
