+++ tests/events/test_app.py
@@ -18,7 +18,7 @@ from tiferet.events.app import (
     RemoveServiceDependency,
     RemoveAppInterface,
 )
-from tiferet.events.settings import DomainEvent, TiferetError, a
+from tiferet.events.core import DomainEvent, TiferetError, a
 from tiferet.domain import (
     AppInterface,
     AppServiceDependency,
