+++ tests/mappers/test_core.py
@@ -1,6 +1,9 @@
-"""Mapper Settings Tests"""
+"""Tiferet Mapper Core Tests"""
 
 # *** imports
+
+# ** core
+from typing import Any, ClassVar, Dict
 
 # ** infra
 import pytest
@@ -8,7 +11,7 @@ from pydantic import Field, ValidationError
 
 # ** app
 from tiferet.domain import DomainObject
-from tiferet.mappers.settings import Aggregate, TransferObject
+from tiferet.mappers.core import Aggregate, TransferObject
 from tiferet.assets import TiferetError
 
 # *** fixtures
