+++ tests/utils/test_file.py
@@ -8,7 +8,7 @@ import pytest
 # ** app
 from tiferet.utils.file import FileLoader
 from tiferet.events import a
-from tiferet.events.settings import TiferetError
+from tiferet.events.core import TiferetError
 
 # *** fixtures
