+++ tests/utils/test_sqlite.py
@@ -13,7 +13,7 @@ import pytest
 # ** app
 from tiferet.utils.sqlite import SqliteClient
 from tiferet.events import a
-from tiferet.events.settings import TiferetError
+from tiferet.events.core import TiferetError
 
 # *** fixtures
