+++ tiferet/utils/json.py
@@ -11,7 +11,7 @@ import json
 # ** app
 from .file import FileLoader
 from ..events import RaiseError, a
-from ..events.settings import TiferetError
+from ..events.core import TiferetError
 
 # *** utils
